refactor(wave3-step2): mechanically split monitor/trace impl behind facades
